(feature) API endpoint to accept a SubmissionEntry without a file
Most of the time we'll be ingesting uploaded files and splitting them
into separate entries.

This endpoint, however, allows an entry to be created without it being
associated to a file. This will be used in the eventual case where
suppliers can choose to post their returns directly to a public-facing
Data Submissions Service API.