(feature) Tasks can be filtered by user_id
To do this we first get a list of all `Suppliers` for a given
`user_id`, via the `Membership` model. We then filter out tasks that are
assigned to these suppliers.

Eventually this will probably be more complex - i.e. show user their
tasks AND the supplier's unassigned tasks, but this is a good start.