Update seeds.rb for local testing
This seeds the database with a single supplier that has multiple tasks
in various states:

- An unstarted task with no submissions
- A started task with the state "in_progress" (submission created but no
file/ingest started) started)
- A started task with the state "processing" (a submission has been made
and the file is being processed)
- A started task with a valid submission (all rows have been processed
and are valud)
- A started task with an invalid submission (all rows have been
processed and some have errors)
- A successfully completed task (the submission has been validated and
marked as completed)

Note that these seeds represent the idealised state of the world, but
not necessarily the state that the system currently puts the submissions
into. for example, there is no part of the system that currently
transitions a "in_progress" task to "processing" when ingest starts. Our
next task is to make sure the system performs the transitions as
required so that the task and submission always have an up-to-date
state.