Update entries API to accept multiple errors
The lambda sends an array of validation errors when it finds errors.
This updates the API to accept the validation errors in the format they
are being sent in.

Note that the strong params gets complicated here because we are having
to accept an array of hashes, and Rails assumes that this will be for a
nested model. This isn't viable longer term where we may need to set
different keys on the validation errors. At that point it will probably
make sense to whitelist the validation error params separately to strong
params and assign them directly.