(feature) Allow users to be imported from a CSV
As this is designed to be run from the rails console, `UserImport.new`
expects an instance of Auth0Client as its second parameter.

It will then create a new Auth0 user for each email address included
in the CSV file, skipping any that already exists. Their passwords and
uid will be printed to the console.