Transition errored submissions to "in_review"
If all the rows have been processed but some have errors, we want to
transition the submission to an "in_review" state so that the user will
see the errors and is able to correct them on the frontend.

Note that we're allowing transition from "pending", as well as
"processed". In theory, we should only ever transition a "processing"
submission to "in_review", but at the moment there is no part of the
system that transitions a submission out of "pending"; instead the
ingest and validation is triggered whilst they are still "pending". The
plan is to update the ingest script so that it transitions the
submission to "processing" when it starts, but until that is done we
wan to be able to see the submissions flowing through the various
states.