Add Customer model
This is the model that represents the part of government a supplier is
selling to. This model has become necessary so that we can generate the
report for finance that groups invoice totals and levys according to the
customers' sectors, i.e. "central government", or "wider public sector".

I'm keeping the attributes on this model to a minimum for now to avoid
having to add code we don't need just yet. Right now, all we really need
are the URN number and the sector. That said, I've additionally included
the name and postcode, mainly because we are expecting to need these two
fields very soon when we add the URN look-up tool to the supplier-facing
application.

I've also avoided adding too many validations or constraints on fields
such as the URN number. We can define and implement those when we get to
the backend administrative part of the system.