Add script for importing customers
Can be used to seed the database with all the customers we currently
know about:

  $ rails runner db/data_migrate/20180807095900_import_customers.rb

The CSV data was generated using the URNs file we are currently sharing
with suppliers. Some of the customers from that file have been excluded:

  - Customers with "Active" set to "False" as business cannot be
  reported against these
  - A handful of "bucket code" URNS that shouldn't be reported against
  by suppliers (e.g. Education/991100)

Note that some customers have been given a postcode of 'XXXX'. I'm
assuming this was to get around some limitation in MISO that meant a
postcode needs to be set. Rather than carry this over to the new system,
the script will ignore these postcodes.

Also note that I've updated the rubocop config to exclude the
Rails/Output rule for the data migrations, as there is no reason to
prevent outputting directly to STDOUT from scripts like this that are
expected to be run directly from the terminal.