Introduce Reports::CodaFinanceReportRow
This will be used to generate the report for Coda (the finance system)
summarising the invoicing details for the daily submissions. This
version summarise all the entries. The final report will need to break
down summaries by sector. That's coming in a future commit.

Note that this version hardcodes the sheet name, totals column name, and
levy amount to the three specific frameworks we currently support:
RM3756, RM3786, and RM3787. We will look to abstract this out at the
point where we have to support frameworks with different
characteristics.