Reports::CodaFinanceReportRow scopes to sector
The report requires that submission summaries be reported by sector.

Note that 'Customer URN' column name has been hardcoded into the
SubmissionEntry#sector scope. This is OK for now whilst we only have the
three Frameworks that all share the same column name, but will need
resolving at the point we add more frameworks. It probably makes sense
at that point to introduce an actual `urn` attribute on the
SubmissionEntry model and set this at the point of ingest.