Add CodaFinanceReport for reporting submissions
This will be used to generate CSV reports for completed submission to be
fed into the current finance system, Coda. For now we're generating this
report manually, thus it outputs as a string to make it easy to run the
report and copy the output from the console.

Eventually this report will need to be generated on a schedule daily,
with the resulting file put somewhere for CCS to perform some automatic
processing and importing.