Configure development ActiveStorage to use S3
The development environment is configured to use Amazon S3 storage so
that the interaction between the supplier app and the API can be
performed locally.

Note: the following environment variables will need to be configured
for this to work:

  AWS_ACCESS_KEY_ID
  AWS_SECRET_ACCESS_KEY
  AWS_S3_REGION
  AWS_S3_BUCKET

Define these in a .env or .env.development file to make use of the
dotenv-rails gem installed with the app.