Endpoint for attaching file blobs on submissions
This endpoint will be used by the supplier-facing application to tell
the backend API where it can find the uploaded submission file.

The supplier-facing application handles the mechanics of uploading
user's files to S3 using ActiveStorage. By sending the details of the
ActiveStorage::Blob to the backend API, the API can create it's own
record of the file and thus make it available for the system to access
and manipulate as required.

This need to sync low-level file storage details using the API in this
way is a bit of a smell, but given the architecture as it has been
defined right now this feels like the most pragmatic way to make it so
that the supplier front-end can handle the file upload whilst still
having the backend API as the authoritative record of the file itself
without the backend having to read the file once it's been uploaded.