Make Export::Tasks real
Export::Tasks is responsible for taking a list of tasks and an IO-like
output (responding only to #puts) and writing to that output:

- a header
- one CSV line per Task via a transform in
  `Export::Tasks::Row#to_csv_line`

This lets us keep the generation simple and the transform separate while
allowing for output to STDOUT, File or StringIO as the situation
demands.

The rake task remains a bit of fluff which still has responsibilities
beyond its station (opening and naming the file to write being things
that need eventually to go somewhere else).