Add and expose SubmissionEntry#entry_type column
This will be used to identify the type of entry, i.e. whether it's an
"invoice" or an "order" row. We want this to be a first-class attribute
of the model as we will often need to scope by the type of entry,
specifically in things like the finance report and data exports.

Note that we'll soon start importing other types of entry (e.g. bids)
and so the list of valid entry types will need to be updated, but a
decision on that should be made later when we know which additional
types of entry we will be importing and whether they need to be
individually identifiable or not (i.e. do we add "bids" to the list or a
catch-all "other").