Simplify submission entry factory (no file)
Strictly speaking a submission entry doesn't need a file to be valid.
Removing this will greatly reduce the number of dependent
records/objects created when setting up multiple submission entries
(i.e one submission file per-entry). If a file is needed, then it should
probably be created explicitly and linked to the associated entries and
submission.