Make submission entry factory types more explicit
Right now (and for perhaps for the foreseeable), submission entries
should only ever be an "invoice" or an "order". This reworks the
factories and their use to make this more explicit. At the point we
actually need to start ingesting data for sheets other than orders or
invoices we can adjust accordingly.