Rename levy column to management_charge
The "levy" term is deprecated, with "management charge" being the
correct term to describe the charge.

Note that because we are between submission periods and are not
expecting any submissions until next month, we can roll this out as a
single release and update the calculate lambda in our own good time
without having to worry about requests being made against the API using
the wrong parameter names.