Add SubmissionEntry.ordered_by_row scope
This will be useful when constructing the data structure that will show
validation errors to the user, where we want to show them the errors in
row-order, making it easier to correct them.

Note that the clause string passed to the order() method is being
wrapped in Arel.sql() to kill a "Dangerous query method" deprecation
warning because non-attribute arguments will be disallowed in Rails 6.0.