Expose Submission#sheet_names
We want to know the sheets that a submission has entries for so that we
can query against each sheet's entries separately, specifically to grav
the first few errored rows to display to the user.

Note that the clause string passed to the pluck() method is being
wrapped in Arel.sql() to kill a "Dangerous query method" deprecation
warning because non-attribute arguments will be disallowed in Rails 6.0.