Make Export::CodaFinanceReport use ToIo/CsvRow
Makes the shape of the Coda finance export the same shape as the other
exports. Coda outputs two rows for every model in the relation, so
define ToIo#output_row and let it be overridden.

Slight downside: in losing Row#data we lose the (quite useful) hash
mapping. To ameliorate, use the same method names called from
CodaFinanceReport::Row#row_values in the public interface and delegate
these to private methods.