Data migration to create October Suppliers
As well as adding the suppliers to the system, this data migration will
output a hash mapping supplier names to the IDs they have been created
against. This can be used in the corresponding script to create the
users for these suppliers against Auth0 and the frontend application.

Coda references taken from Excel file provided by James Ryan.

Run with:

  rails runner db/data_migrate/20180924143116_add_october_suppliers.rb