Make Export::Submissions::Extract count IDs
If we use COUNT(DISTINCT orders), PostgreSQL will, by default, compare
a hash of the whole row.

If we use COUNT(DISTINCT orders.id), PostgreSQL will compare the
identity of the underlying submission entry.

The difference in performance tested against a mid-Sep cut of production
data is ~5s -> 150ms.

Comparative query plans are at

https://gist.github.com/rgarner/4276e198f084e1574505b9bd9e56f0be

The differences are subtle; check the `width` cost of the comparisons
on these two lines:

Whole record:
https://gist.github.com/rgarner/4276e198f084e1574505b9bd9e56f0be#file-count_distinct_whole_record-sql-L47

Ids only:
https://gist.github.com/rgarner/4276e198f084e1574505b9bd9e56f0be#file-count_distinct_ids-sql-L46