Add Utility class for generating monthly tasks
Every month suppliers need to make submissions against the frameworks
for which they have agreements in place for. This code generates the
records that represent that monthly task.

For now the due date is fixed at seven days after the subsequent month.
This may need to be more dynamic in future as some frameworks may have
different contractual terms, for example five working days instead of
seven calendar days.