Add rake task for generating monthly tasks
For now the rake task requires that the period be specified, for
example:

  rake generate:tasks[8,2018]

It will probably make sense to have the rake task create the tasks for
the current period at some point so that it can easily be run on an
automated schedule, but this will do for now.

Also disables the Andor rubocop as it's too agressive - the use of the
and/or operators for control flow is legit in my ([and others]) opinion.

[and others]: http://www.virtuouscode.com/2010/08/02/using-and-and-or-in-ruby/