export:contracts map contract start/end dates
A compromise decision is being made here to keep things simple.
We're lumping mapped contract fields in with mapped invoice fields,
on the basis that a) it's simple and b) there aren't any collisions
yet. We expect that to change when the Product* fields are mapped
and at that point template.rb will have to acknowledge the changes.

A second and larger problem is that we are passing dates through
un-transformed. They appear to be coming in M/DD/YY form on this
particular example record. That isn't the case for everything.

We need a much bigger discussion about dates.