rake export:contracts Additional8 fields
Slightly contentious commit – the first field, Additional1, is still
mapped to Matter Name, because it's the same (as far as I can tell with
incomplete information) as Invoices.

The rest of the fields are blank, as they don't map to anything (yet)
and it's perfectly valid for them to be blank as that's the default
behaviour of AdditionalN fields. That does mean we can't use #MISSING
to point up that these aren't mapped yet.

Also, a refactoring: there's no real need to define 8 methods on the
SubmissionEntryRow, as we tend to take these things in one go,
so use *values_for_additional to blit them onto the row_values of
invoices and contracts respectively in one go.