Update customer filterer to allow inactive
We need to include inative customers in the system because suppliers may
make a submission including invoicing against customers that are no
longer active, but were at the time of invoicing. This updates the
filter utility to import both active and inactive customers.

This change also necessitated the broadening of the filtering of
"internal" customers as the existing regular expression was not enough
to filter out the inactive ones.