Temporarily bypass management charge calculation
We've realised that we [removed an endpoint the calculator relies] on to
perform the calculation, which has broken calculation! For now we're
going to skip calculation so that users are not blocked from completing
their submissions.

We will need to re-instate calculation and perform calculation against
October submissions before we can generate the finance and date
warehouse exports.

[removed the endpoint that the calculator relies](https://github.com/dxw/DataSubmissionServiceAPI/commit/bdd0a193e59e192f2710d90a494b7add09c82db5)