Prevent duplicate SubmissionEntries from being created
When the ingest lambda fails, it automatically restarts which was
generating duplicate entries for the same row/sheet in a given
spreadsheet.

Now we check whether the `submission_id` and `source` already exist,
and return an HTTP 204 No Content status if this is the case (I
originally chose 409 Conflict, but this may require changes in the
lambda).