Spec the generators
You can and should argue that this is backwards – that the previous
commit has established a generator before its tests exist. Mea culpa –
I have retrofitted these because the RSpec/generator testing universe
is something of a backwater. The standard rails generator testing
toolkit is unashamedly MiniTest-based.

This commit is an effort to make those two worlds talk to one another a
little; spec/support/generators says that any spec with a
type: :generator will have some extra MiniTest-focussed modules mixed in
on it, and that before every test, a directory that you specify in
an RSpec `destination` let will be cleared down ready for the output
of the generator.

In our case our generator generates precisely one file, so many of the
special setups needed for testing generators don't apply. What's in
spec/support/generators supports our use case and is a mini translation
of Rails::Generators::Testing::SetupAndTeardown.

We haven't specified any RSpec-specific matchers (although we can use
expect(definition).to match in most of our assertions), so everything
hinges on Rails::Generators::Testing::Assertions#assert_file, which can
take a block. So we use that block to populate our subject(:definition)
using @definition, which is a bit of a sleight of hand.

Anyway, I hope you agree it reads well enough and does a job.