For our exporter specs, failures are truncated
It's not useful to see a long line with '...' instead of the source of
the failure in the middle, so for any spec in spec/lib/tasks/export,
add an `metadata[:exporter_spec] = true` definition.

Use that to include a Concern that temporarily extends
RSpec's max_formatted_output_length for the duration of the example.