Require framework definitions up-front
Export::SheetMappings requires the framework definitions to be present.
Rather than loading them on-demand with require_dependency or leaving it
to Rails autoload (which calls String#underscore behind the scenes and
hence requires the file to be called Rm1234.rb and the class Rm1234),
enumerate the files in app/models/framework/definitions (excepting base)
and require them separately once.