Add some temporary code for reporting spend/charge
We will use this rake task to report the monthly spend and management
charge in the interim until we have a proper administrative interface
that gives the figures.

Note that until we have both the total value and management charge
stored as database columns on submission_entries, calculating the total
spend and management charge is convoluted, inefficient and hacky. Things
get much simpler once the work to normalise the data and model the
framework definitions in the Ruby application is complete.