Backfill submission_entries.total_value
Uses a list of expected total value fields for Sep/Oct and looks through
them one by one with COALESCE before stripping £, comma and space and
CASTing to `decimal`. It corrects for the values in some orders that
would break the CAST with a CASE statement that explicitly deals with
those values.

See the comments for further context.