Update Export::Submissions::Extract with totals
Now we have SubmissionEntry#total_value, use it in the submissions
export. This shows our previous LEFT JOINs with COUNT(DISTINCT) to have
hidden a cartesian product. Pushing the COUNTs into JOINed subqueries
and out of the SELECT lets us move the counting and summing to the
appropriate place such that DISTINCT isn't required and we don't create
a accidental cross-join.

Because of this the subqueries can legitimately produce NULL as invoice
and entry counts, so COALESCE these to 0 for no-business submissions.