Add management charge rate to framework definition
For now we can model this as a primitive number representing the
percentage for the management charge rate. We expect that this will
morph into some kind of strategy-pattern style approach when we are
looking to support frameworks with more complex management charge
calculations (for example those with discounts or lot-based
calculations).