Format known date values in data warehouse export
At the moment, dates are being ingested by the lambda-based ingest
process as strings in consistently inconsistent way: they either appear
as "MM/DD/YY" or "DD/MM/YYYY"*. The export needs to report them in
ISO8601 format. This makes that the case for the date fields we have
known sheet mappings for.

Note that this does not cover all cases where we have ingested a date as
there are other date fields that sometimes map to the additional fields.
For example RM3710 maps 'Lease Start Date' to 'Additional8' in the
export. To handle these cases we'll need a more sophisticated solution
that introspects on the field definitions in the framework definitions
to determine when the value should be formatted as a date, but that's a
more involved piece of work that we don't have time for right now (there
is a pressure to deliver the export asap so business/finance can perform
the necessary checks and invoicing). Once we start the work to perform
validations in Ruby (leaning on ActiveModel), we'll have the necessary
plumbing in place to determine if a given field is a date and can at
that point encapsulate the formatting so it happens automatically inside
`Export::SubmissionEntryRow#value_for` based on each framework's
definition instead of explicitly in `Export::Invoices::Row` and
`Export::Contracts::Row`.

Additionally/alternatively we expect to fix/replace ingest such that
dates are ingested and validated in a consistent format that may make
this redundant.

* mostly. Turns out there are actually a couple of instances in the
production data that don't, hence the rescuing ArgumentError to handle
invalid dates.