Fix RM3767 framework definition mismatches
Both 'Total Cost (ex VAT)' and 'VAT Amount Charged' had case that didn't
match the case in the template file or the case used in the current
lambda-based ingest process, which was resulting in'#NODATA' being
reported in the data warehouse export.

It's also worth noting that although the definition contains a 'Customer
PostCode' field, the template doesn't have such a column. It appears
that at some point the column was removed from the template. We're
leaving it in place until some further investigation is carried out to
decide what's best to do.