Only backfill total_value on validated entries
When this data migration was run on production it failed because their
are entries in the production data that do not match the expected format
for the totals fields, for example: "(17.60)" and "$£48.46". The safest
way to ensure this backfill will always work is to only perform the
backfill on entries that have been validated, thus ensuring they will
match the expected format.