Add submission_entries#management_charge column
We're moving towards a per-entry management charge calculation as a)
there are frameworks where the calculation varies per-row, and b)
the management charge needs to be rounded down, and calculating the
charge against the total would result in a significant difference in the
charge for submissions with large numbers of entries.

Note that we're storing this value to 4 decimal places. This is to
replicated the existing behaviour in the current system (MISO).

Co-authored-by: leeky <leeky@dxw.com>