Add management charge calculation to definitions
This feels like it's probably the right place to put this for now, but
I'm not entirely comfortable with the location/organisation of the spec.
Will need to be addressed sooner rather than later.

Note that the calculation is being performed to 4 decimal places to
replicate the behaviour in the existing system (MISO).