Add missing inactive customers
It looks like https://github.com/dxw/DataSubmissionServiceAPI/pull/80
missed the mark slightly, as a bunch of inactive customers were not
included in the data migration.

This was noticed because the coda finance report was reporting numbers
that were slightly at odds to the managment charge totals for the entire
submission. The report relies on the SubmissionEntry#central_government
and SubmissionEntry#wider_public_sector scopes to grab entries for each
sector. The problem is, if the entry has a customer URN that isn't in
the database it won't appear under either scope, resulting in it not
appearing in either calculation, thus the under reporting.

We need to make sure that the ingest/validation process is tightened up
such that this isn't possible in the future.