Add SubmissionEntry association with Customer
This uses the Customer's URN as the foreign key, which is the unique
identifier used by CCS and suppliers. The foreign key constraint in the
migration will help maintain data integrity, as any attempts to change
the URN of a customer or delete a customer associated with an entry will
raise a foreign key violation error.