Trigger submission management charge calculation
Whilst we're still half-in/half-out of lambda land, we will trigger the
management charge calculation at the point the last submission entry has
been validated (and only if all the entries are valid).

Note that the SubmissionStatusUpdate class should almost certainly not
exist and the transition and firing of the various events should
probably be encoded in the state machine on Submission. Putting off that
work until we are fully in Ruby-land for ingest, validation and
calculation.