Data migration for users with multiple suppliers
This is to work around an issue where some users have access to more
than one supplier's tasks. To make it so that these users are able to
distinguish tasks between their suppliers, we set the tasks'
description field to the supplier's name so that the frontend
application can show it when displaying the task. Note we are only doing
this for the subset of suppliers that share a user.