Update RM1070 "VAT inc" validation to match data
Although the framework_miso_fields.csv file specifies this field as
being "System:Boolean", we are ingesting it as a string and validating
it as either 'Y' or 'N'. For now we can avoid any type-casting for this
field and simply validate against the expected string values.

Note that this introduces a specific spec file for the specific
framework definition file for RM1070 so that we can verify the
validations behave as expected against the data we have already ingested
and validated/invalidated. Until such a time that we have abstracted
away common fields and their validations, this is the safest way to
ensure we have framework definitions that behave as we need.