Update RM1070 date validation to match data
Because of certain quirks of the current lambda-based ingest we are
currently seeing dates in two different string formats – either as
"DD/MM/YYYY" or "MM/DD/YY" – and the existing lambda-based validator
currently validates both.

To make sure the Ruby-based validation is consistent with the
lambda-based one, we introduce a new validator that recognises the
two formats and validates them as long as they actually
correspond to a valid date.