Don't cast UNSPSC/Quantity in framework definition
Although we don't validate these at present, we should be (these were
both fields what we relaxed validations for because of the issue
discovered in MISO where certain validations were not being performed).
To make it less likely we'll hit issues later if/when we re-instate
validations for these fields we should avoid casting them to integers as
that results in non-numeric string values being cast to 0.