Add support for API to return supplier name
The API returns the supplier id but not the supplier name.

Users who complete tasks for multiple suppliers need to know which
supplier a given task is for, we want to show the supplier name on
the task view, we hacked this by using the redundant Task.description
to store the supplier name.

This adds support in the API to return the supplier name as
Task.supplier_name.

The next step will be to know when to display the supplier name based
on whether the current logged in user is linked to multiple suppliers.