Refine User.search results
Fixes two bugs:

- Users that weren't linked to a supplier were being excluded from the
results becauces of the use of `joins`
- Users that were linked to more than one supplier would be returned
multipe times, also because of the use of `joins`

Switching to `eager_load` forces ActiveRecord to left outer join to the
suppliers table so all users are returned, even those that are not
associated with a supplier. This also resolves the issue of duplicate
users being returned when a matching user belongs to more than one
supplier.