Mark RM3710 lease start/end dates as optional
These fields are not mandatory according to the template* and were not
required in the previous JS-based version of the validator.

* The wording in the template actually says "MANDATORY FIELD - THIS
FIELD MUST BE COMPLETED BEFORE SUBMITTING Lease Start Date Please enter
the start date of the lease IF THIS FIELD IS NOT APPPLICABLE, PLEASE
ENTER 01/01/1999" but we're going to change that so it's clearer it can
be left blank.