Add supplier search when linking to users
To make it easier for support staff to find the correct supplier when
linking them to a user we add a basic search feature. We also display
the suppliers sorted alphabetically.

Note that the use of a LIKE query with a leading wildcard is not going
to be particularly efficient as PostgreSQL wouldn't be able to make use
of an index even if we had one. If performance becomes an issue in the
future we should consider a dedicated search system, or possibly install
the pg_trgm extension to add index support for string matching.