Tweak Framework factory to avoid short_name clash
This is resulting in sporadic test failures.

Because there are some specs that test against the actual framework
definitions and thus need to create framework records with matching
short names such as "RM1070", we can end up in a situation where the
factory attempts to creates a framework with a clashing short name
(e.g. when the sequence number has reached 70 in a spec that also
creates a framework with the explicit short_name of "RM1070"). By
changing the format to one that can't match any existing frameworks (at
least any that we know of) we avoid these flakey tests.