Push entry data hash validation down to model
We're about to add additional complexity to the validation of the
submission entry data by introducing lot-based validation. By pushing
the validation of the entry and its data hash into the SubmissionEntry
model, where it's encapsulated better and makes it easier to interogate
the relations of the submission entry, i.e. the lots the supplier is
allowed to submit against.