EntryData initialization takes an entry instead of data
We will need the EntryData class to know about its entry for lot base validations.

We call super EntryData#initialize so that ActiveModel can set up
the attributes for the instance. We have to call it with the empty
paranthesis because otherwise Ruby by default passes on the passed-in
entry parameter to the overridden method, which isn't what ActiveModel
expects.

Co-Authored-By: Tekin Suleyman <tekin@tekin.co.uk>