Ensure duplicate agreements cannot be created
Now that we are about to model and validate against lots, we need to
should make sure each supplier has a single unique agreement against any
given framework to avoid any complications or errors. This may change in
the future when we introduce the dynamic frameworks that suppliers
can join and leave at will. At that point we'll likely introduce
start/end dates for agreements and this can switch to validating there
is no overlap in dates instead.