Add BankHolidays class to fetch bank holiday dates
We'll use this to adjust the due date on a submission window based on
the presence of bank holidays. The JSON data published on GOV.UK
includes past and future dates. We can cache the data for a fair period
because it changes infrequently, and normally a couple of years in
advanced.