Reject API requests without X-Auth-Id header by default
Require all API requests to include the `X-Auth-Id` header which
contains the auth_id of the currently logged-in user, unless the
endpoint is called by the ingest lambda (which doesn't know about the
current user).

Exclude submission_entries controller from this authentication as it
is used by the ingest lambdas, which do not have access to this
information

NB: When I added the `headers: {}` hash to the HTTP requests within
`spec/requests`, Rubocop started complaining about positional arguments.

I believe this to be a false positive, as letting RuboCop correct the
errors results in multiple test failures - so I've excluded the request
spec from being checked against the `Rails/HttpPositionalArguments`
rule.