(fix) Ensure generated temp pwd is auth0 valid
Previously, user creation would occasionally fail because
the random generated password didn’t contain all the
required types of characters for auth0.

This password is only used to seed the user account
and will be reset by the user upon their first login.

As such, this commit adds a minimal set of characters
to ensure that auth0 will accept the password.