make "no business" API endpoint idompotent
Some users were double clicking the button to submit a no business submission, which resulted in a 500 error when we tried to mark a task as completed for the second time.

This commit will ensure that you can call no_business on a task, and it will not create a new submission or try and update its status, instead returning a not_modified HTTP response.