(chore) Configure new cop to ignore framework defs
The Rubocop update brought with it a new cop,
`Naming/FileName`, enabled by default.

This cop finds the framework definition filenames guilty
of not being in snake_case.

We already exclude these classes from the existing cop
`Naming/ClassAndModuleCamelCase`, so there is a good
precedent for excluding them from the new one, too.