(chore) Ignore alignment of hashes
A new cop enabled by default is asking us to have an
enforced style of alignment for hashes. Our existing code
has a mix of styles.

James D and I discussed and we don’t particularly care,
so I chose to disable this cop for now, which is no
different from what we had before the rubocop upgrade.