(feature) Validator for interdependent fields
Previously, we were only validating the inclusion of a value
such as Primary Specialism in the set of accepted values for all
primary specialisms across service types, core and non-core.

This commit introduces a custom validator that allows us to check
the inclusion of a value in the restricted set of values for the
higher order category of that specific submission (in the case
of Primary Specialism, the Service Type determines the valid
values; other frameworks have a Lot Number that defines what
services can be provided).

The validator depends on defining a mapping linking the "parent"
field with the set of acceptable values of the dependent field.