Add periodic job to publish queue size to AWS
Uses the sidekiq-cron gem to queue a job that will publish the queue
size to AWS every thirty seconds.

sidekiq-cron runs a thread alongside Sidekiq workers to schedule jobs at
specified times. It is safe to use with multiple sidekiq processes or
nodes as it uses a Redis sorted set to determine that only the first
process that asks can enqueue scheduled jobs into the queue.

Note that the cron syntax in config/sidekiq_schedule.yml is slightly
different to standard cron in that it allows per-second scheduling.
Because we schedule the job to run every thirty seconds, the polling
interval needs to be reduced from the default of thirty seconds.