(fix) No business endpoint returns the submission
Previously, we were attempting to respond to a duplicate
`no_business` submission request with a status of ‘Not
Modified’ *and* send the submission in the body.

This was resulting in an error on the supplier facing app,
because Not Modified automatically returns an empty body.

This error was triggered by the user clicking the submission
button too quickly twice in a row, which fired off two requests.

The first request would result in a 201 Created status, while
the second would find the task already completed.

While ‘Not Modified’ would technically be the correct status
for this situation, we considered it more important that the
body be returned in the response, so the supplier gets
redirected to their completed task view.