Introduce Workday::SubmitCustomerInvoiceRequest
This generates the SOAP API request payload to create an invoice in
Workday for a given submission.

The Workday API requires that a revenue category and cost center is
specified in the requests. However we do not currently have a mechanism
to identify these, so they are hardcoded for the time being. A new API
endpoint is being added to Workday that will allow us to map a framework
to its revenue category and cost center.

Note that the authentication headers are not yet included, those will
come in a subsequent commit.